requirements.txt

streamlit
shodan
requests
whois
pandas
pillow
spacy
python-whois
pyexiftool
